requirements.txt
absl-py==2.3.0
agno==1.7.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.38.42
botocore==1.38.42
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
curl-cffi==0.11.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.16
duckduckgo-search==8.0.5
email-validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.4
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-genai==1.24.0
googlesearch-python==1.3.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.2.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
lancedb==0.24.0
lightning==2.5.2
lightning-cloud==0.5.70
lightning-sdk==0.2.18
lightning-utilities==0.14.3
litdata==0.2.45
litserve==0.2.12
lxml==6.0.0
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.5.0
multitasking==0.0.11
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
oauthlib==3.3.1
ollama==0.5.1
openai==1.93.0
overrides==7.7.0
packaging==25.0
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pgvector==0.4.1
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
primp==0.15.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==4.23.4
psutil==7.0.0
psycopg==3.2.9
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.3
pypdf==5.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.8
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.3.2
scipy==1.11.4
send2trash==1.8.3
setuptools==78.1.1
shellingham==1.5.4
simple-term-menu==1.6.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tantivy==0.24.0
tenacity==8.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tomli==2.2.1
torch==2.7.0+cu128
torchmetrics==1.3.1
torchvision==0.22.0+cu128
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
wget==3.2
wheel==0.45.1
widgetsnbextension==4.0.14
yarl==1.20.1
yfinance==0.2.64
youtube-transcript-api==1.1.1
